+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1 under development
 
 - Enh #225: Typecast refactoring (@Tigrov)
+- Enh #226: Add support for auto increment in primary key column. (@terabytesoftw)
 
 ## 1.0.0 April 12, 2023